Wrapper Function for Double Machine Learning with Cross-Fitting

Implements double machine learning inference with k-fold cross-fitting (default).

DML(data, y, d, model = "plr", k = 2, S = 1, smpls = NULL,
  mlmethod, dml_procedure = "dml2", params = list(params_m = list(),
  params_g = list()), inf_model = "IV-type", se_type = "ls",
  aggreg_median = TRUE, ...)

Arguments

data
Data frame.

y
Name of outcome variable. The variable must be included in data.

d
Name of treatment variables for which inference should be performed.

model
Inference model to be implemented, e.g. partially linear regression (plr), partially linear instrumental variable regression (plriv) (default).

k
Number of folds for k-fold cross-fitting (default k=2).

S
number of repetitions (default S=1.

mlmethod
If multiple target coefficients are provided and different mlmethods chosen for each coefficient, the names of the methods are required to match. A list of available methods is available at https://mlr.mlr-org.com/articles/tutorial/integrated_learners.html.

dml_procedure
Double machine learning algorithm to be used, either "dml1" or "dml2" (default).

params
If multiple target coefficients are provided and different mlmethods chosen for each coefficient, the names of the methods are required to match.

inf_model
Inference model for final estimation, default "IV-type" (...)

se_type
Method to estimate standard errors. Default "ls" to estimate usual standard error from least squares regression of residuals. Alternatively, specify "IV-type" or "DML2018" to obtain standard errors that correspond to the specified inf_model. The options chosen for inf_model and se_type are required to match.

aggreg_median
logical indicating how target estimators and standard errors should be aggregated for repeated cross-fitting. By default (TRUE), the median over all estimators as taken as the final estimator. If FALSE, the mean is calculated over all estimators.

...
further options passed to underlying functions.

Value

Result object of class DML with estimated coefficient and standard errors.