Bootstrap Implementation for Partially Linear Regression Model

Multiplier bootstrap to construct simultaneous confidence bands for multiple target coefficients in a partially linear regression model (PLR) with double machine learning.

bootstrap_plr(theta, d, u_hat, v_hat, v_hatd, inf_model, se, bootstrap,
  nRep)

Arguments

theta
final dml estimator for the partially linear model.

d
treatment variable.

u_hat
Residuals from \(y-g(x)\).

v_hat
Residuals from \(d-m(x)\).

v_hatd
Product of v_hat with d.

inf_model
Inference model for final estimation, default "IV-type" (...)

se
Estimated standard error from DML procedure.

bootstrap
Choice for implementation of multiplier bootstrap, can be set to "normal" (by default), "none", "Bayes", "wild".

nRep
Number of repetitions for multiplier bootstrap, by default nRep=500.

Value

List with bootstrapped standard errors (boot_se) and bootstrapped coefficients.