Double Machine Learning for Interactive Regression Model.

Double Machine Learning for Interactive Regression Model.

dml_irm(data, y, d, k = 2, smpls = NULL, mlmethod,
  params = list(params_m = list(), params_g = list()),
  dml_procedure = "dml2", inf_model = "ATE", se_type = "ATE",
  bootstrap = "normal", nRep = 500, ...)

Arguments

data
Data frame.

y
Name of outcome variable. The variable must be included in data.

d
Name of treatment variables for which inference should be performed.

k
Number of folds for k-fold cross-fitting (default k=2).

mlmethod
List with classification or regression methods according to naming convention of the mlr package. Set mlmethod_g for classification or regression method according to naming convention of the mlr package for regression of y on X (nuisance part g). Set mlmethod_m for classification or regression method for regression of d on X (nuisance part m).

params
Hyperparameters to be passed to classification or regression method. Set hyperparameters params_g1 for predictions of nuisance part g1, params_g0 for nuisance part g0, and params_m for nuisance m.

dml_procedure
Double machine learning algorithm to be used, either "dml1" or "dml2" (default).

inf_model
Estimator for final estimation, default average treatment effect "ATE". Alternatively switch to "ATET" for average treatment effect on the treated.

se_type
Method to estimate standard errors (redundant / identical to inf_model).

bootstrap
Choice for implementation of multiplier bootstrap, can be set to "normal" (by default), "none", "Bayes", "wild".

nRep
Number of repetitions for multiplier bootstrap, by default nRep=500.

...
further options passed to underlying functions.

resampling
Resampling scheme for cross-fitting of class ResamplingCV.

Value

Result object with estimated coefficient and standard errors.