Multiple Testing Adjustment of p-values for S3 objects DML

Multiple hypotheses testing adjustment of p-values for double machine learning.

p_adjust(x, ...)

# S3 method for DML
p_adjust(x, method = "RW", ...)

Arguments

x
an object of S3 class DML.

...
further arguments passed on to methods.

method
the method of p-value adjustment for multiple testing. Romano-Wolf stepdown ('RW') is chosen by default.

Value

A matrix with the estimated coefficients and the p-values that are adjusted according to the specified method.

Details

Multiple testing adjustment is performed for S3 objects of class DML. Implemented methods for multiple testing adjustment are Romano-Wolf stepdown 'RW' (default) and the adjustment methods available in the p.adjust function of the stats package, including the Bonferroni, Bonferroni-Holm, and Benjamini-Hochberg corrections, see p.adjust.methods.

Objects of class DML are constructed by DML.

Methods (by class)

DML:

References

J.P. Romano, M. Wolf (2005). Exact and approximate stepdown methods for multiple hypothesis testing. Journal of the American Statistical Association, 100(469), 94-108.

J.P. Romano, M. Wolf (2016). Efficient computation of adjusted p-values for resampling-based stepdown multiple testing. Statistics and Probability Letters, (113), 38-40.